new activity for server nearby connection, debugging
