+++ .gitignore
@@ -13,3 +13,5 @@
 /fakeinput_dist
 app/src/main/res/raw/flyinn_fakeinputlib.jar
 BuildInfo.java
+app/lint-report.html
+server/lint-report.html
