+++ app/src/main/AndroidManifest.xml
@@ -41,17 +41,17 @@
             android:name=".WifiP2PActivity"
             android:label="@string/app_name"
             android:theme="@android:style/Theme.Holo.Light.DarkActionBar">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
         </activity>
         <activity
             android:name=".WebRTCActivity"
             android:label="WEBRTC Example"
             android:theme="@android:style/Theme.Holo.Light.DarkActionBar"></activity>
-        <activity android:name=".MainActivity"></activity>
+        <activity android:name=".MainActivity">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
         <activity
             android:name=".ADBActivity"
             android:label="@string/title_activity_adb"
