+++ server/src/main/AndroidManifest.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.amos.server">
+
     <uses-permission
         android:name="android.permission.ACCESS_COARSE_LOCATION"
         android:required="true" />
@@ -18,6 +19,7 @@
     <uses-permission android:name="android.permission.CAMERA" />
     <uses-permission android:name="android.permission.RECORD_AUDIO" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
+
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
@@ -25,13 +27,8 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@android:style/Theme.NoTitleBar.Fullscreen">
-        <activity android:name=".P2PActivityServer">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
-        </activity>
-        <activity android:name=".WebRTCServerActivity"></activity>
+        <activity android:name=".P2PActivityServer"></activity>
+        <activity android:name=".WebRTCServerActivity" />
         <activity android:name=".MainActivity">
         </activity>
         <activity
@@ -43,11 +40,19 @@
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value="com.amos.server.MainActivity" />
         </activity>
-        <activity android:name=".EventSenderDemo"></activity>
+        <activity android:name=".EventSenderDemo" />
         <activity
             android:name=".BuildInfoActivity"
             android:label="@string/title_activity_build_info"
-            android:theme="@style/AppTheme.NoActionBar"
-            />
+            android:theme="@style/AppTheme.NoActionBar" />
+        <activity
+            android:name=".NearbyServerActivity"
+            android:label="@string/title_activity_nearby_server">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
     </application>
+
 </manifest>
